Rename user profile store
